@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace DeMol2018.BitcoinGame.Domain.Models
 {
@@ -9,7 +7,7 @@ namespace DeMol2018.BitcoinGame.Domain.Models
         public Guid Id { get; set; }
         public int RoundNumber { get; set; }
         public Guid GameId { get; set; }
-        
+
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
 
@@ -17,17 +15,16 @@ namespace DeMol2018.BitcoinGame.Domain.Models
         public bool HasEnded => DateTime.UtcNow > EndTime;
 
         public bool IsActive => HasStarted && !HasEnded;
-        
+
         public Round()
         {
-//            HasStarted = false;
+            Id = Guid.NewGuid();
         }
 
         public void Start(TimeSpan roundLength)
         {
             StartTime = DateTime.UtcNow;
             EndTime = DateTime.UtcNow.Add(roundLength);
-  //          HasStarted = true;
         }
     }
 }
