@@ -1,6 +1,5 @@
 import * as React from 'react';
-import * as BitcoinGameStore from "../store/BitcoinGame";
-import {connect} from "react-redux";
+import {connect, ConnectedProps} from "react-redux";
 import {ApplicationState} from "../store";
 import Login from "./Login";
 import PlayerWallet from "./PlayerWallet";
@@ -8,9 +7,8 @@ import AdminPanel from "./AdminPanel";
 import GameManager from "./GameManager";
 import PlayerResult from "./PlayerResult";
 
-type BitcoinGameProps =
-    BitcoinGameStore.BitcoinGameState
-    & typeof BitcoinGameStore.actionCreators;
+const connector = connect((state: ApplicationState) => state.bitcoinGame);
+type BitcoinGameProps = ConnectedProps<typeof connector>
 
 class BitcoinGame extends React.Component<BitcoinGameProps> {
     public render() {
@@ -33,7 +31,7 @@ class BitcoinGame extends React.Component<BitcoinGameProps> {
             } else {
                 gameContent = <PlayerWallet />;
             }
-        } 
+        }
 
         return (
             <div className="bitcoinGame">
@@ -44,8 +42,4 @@ class BitcoinGame extends React.Component<BitcoinGameProps> {
     }
 }
 
-// Wire up the React component to the Redux store
-export default connect(
-    (state: ApplicationState) => state.bitcoinGame, // Selects which state properties are merged into the component's props
-    BitcoinGameStore.actionCreators // Selects which action creators are merged into the component's props
-)(BitcoinGame);// as typeof BitcoinGameRoot;
+export default connector(BitcoinGame);
