@@ -1,14 +1,12 @@
 import * as React from 'react';
-import { connect } from 'react-redux';
+import {connect, ConnectedProps} from 'react-redux';
 import { ApplicationState }  from '../store';
-import * as BitcoinGameStore from '../store/BitcoinGame';
 import MakeTransaction from "./MakeTransaction";
 
-type WalletProps =
-    BitcoinGameStore.BitcoinGameState
-    & typeof BitcoinGameStore.actionCreators;
+const connector = connect((state: ApplicationState) => state.bitcoinGame);
+type PlayerWalletProps = ConnectedProps<typeof connector>
 
-class PlayerWallet extends React.Component<WalletProps> {
+class PlayerWallet extends React.Component<PlayerWalletProps> {
     public render() {
         return (
             <div className="wallet">
@@ -28,7 +26,4 @@ class PlayerWallet extends React.Component<WalletProps> {
 }
 
 // Wire up the React component to the Redux store
-export default connect(
-    (state: ApplicationState) => state.bitcoinGame, // Selects which state properties are merged into the component's props
-    BitcoinGameStore.actionCreators // Selects which action creators are merged into the component's props
-)(PlayerWallet);// as typeof PlayerWallet;
+export default connector(PlayerWallet);
